requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.6,<0.0.7
boto3==1.28.57
python-magic==0.4.24
pymysql==1.1.0
pymysql[rsa]
mysql-connector-python>=8.0,<9.0